requirements.txt
asttokens==3.0.0
certifi==2025.1.31
charset-normalizer==3.4.1
comm==0.2.2
debugpy==1.8.12
decorator==5.1.1
dnspython==2.7.0
exceptiongroup==1.2.2
executing==2.2.0
idna==3.10
ipykernel==6.29.5
ipython==8.32.0
jedi==0.19.2
jupyter_client==8.6.3
jupyter_core==5.7.2
matplotlib-inline==0.1.7
mysql-connector-python==9.2.0
nest-asyncio==1.6.0
numpy==2.2.2
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
platformdirs==4.3.6
prompt_toolkit==3.0.50
psutil==6.1.1
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.1
pymongo==4.11
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
pyzmq==26.2.1
requests==2.32.3
six==1.17.0
stack-data==0.6.3
tornado==6.4.2
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
wcwidth==0.2.13
